node atualizado com v do tipo double

+++ Node.h
@@ -9,7 +9,7 @@ struct Node {
     int coluna;
     double valor;
 
-    Node(int l, int c, int v,Node* ptrNext = nullptr, Node* ptrBottom = nullptr){
+    Node(int l, int c, double v,Node* ptrNext = nullptr, Node* ptrBottom = nullptr){
         this->linha = l;
         this->coluna = c;
         this->valor = v;
